LPD-69638 refactor: move site/site-ai-hub-rest-client to ai-hub/ai-hub-rest-client
