+++ modules/dxp/apps/ai-hub/ai-hub-rest-client/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay AI Hub REST Client
+Bundle-SymbolicName: com.liferay.ai.hub.rest.client
+Bundle-Version: 1.0.0
