+++ modules/dxp/apps/ai-hub/ai-hub-rest-client/build.gradle
