--- modules/dxp/apps/site/site-ai-hub-rest-client/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay Site AI Hub REST Client
-Bundle-SymbolicName: com.liferay.site.ai.hub.rest.client
-Bundle-Version: 1.0.0
