--- modules/dxp/apps/site/site-ai-hub-rest-client/src/main/java/com/liferay/site/ai/hub/rest/client/function/UnsafeSupplier.java
@@ -1,20 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.site.ai.hub.rest.client.function;
-
-import jakarta.annotation.Generated;
-
-/**
- * @author Feliphe Marinho
- * @generated
- */
-@FunctionalInterface
-@Generated("")
-public interface UnsafeSupplier<T, E extends Throwable> {
-
-	public T get() throws E;
-
-}
