--- modules/dxp/apps/site/site-ai-hub-rest-client/src/main/java/com/liferay/site/ai/hub/rest/client/serdes/v1_0/TaskSerDes.java
@@ -1,260 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.site.ai.hub.rest.client.serdes.v1_0;
-
-import com.liferay.site.ai.hub.rest.client.dto.v1_0.Task;
-import com.liferay.site.ai.hub.rest.client.json.BaseJSONParser;
-
-import jakarta.annotation.Generated;
-
-import java.util.Iterator;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Set;
-import java.util.TreeMap;
-
-/**
- * @author Feliphe Marinho
- * @generated
- */
-@Generated("")
-public class TaskSerDes {
-
-	public static Task toDTO(String json) {
-		TaskJSONParser taskJSONParser = new TaskJSONParser();
-
-		return taskJSONParser.parseToDTO(json);
-	}
-
-	public static Task[] toDTOs(String json) {
-		TaskJSONParser taskJSONParser = new TaskJSONParser();
-
-		return taskJSONParser.parseToDTOs(json);
-	}
-
-	public static String toJSON(Task task) {
-		if (task == null) {
-			return "null";
-		}
-
-		StringBuilder sb = new StringBuilder();
-
-		sb.append("{");
-
-		if (task.getContext() != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"context\": ");
-
-			sb.append(_toJSON(task.getContext()));
-		}
-
-		if (task.getExternalReferenceCode() != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"externalReferenceCode\": ");
-
-			sb.append("\"");
-
-			sb.append(_escape(task.getExternalReferenceCode()));
-
-			sb.append("\"");
-		}
-
-		if (task.getType() != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"type\": ");
-
-			sb.append("\"");
-
-			sb.append(_escape(task.getType()));
-
-			sb.append("\"");
-		}
-
-		sb.append("}");
-
-		return sb.toString();
-	}
-
-	public static Map<String, Object> toMap(String json) {
-		TaskJSONParser taskJSONParser = new TaskJSONParser();
-
-		return taskJSONParser.parseToMap(json);
-	}
-
-	public static Map<String, String> toMap(Task task) {
-		if (task == null) {
-			return null;
-		}
-
-		Map<String, String> map = new TreeMap<>();
-
-		if (task.getContext() == null) {
-			map.put("context", null);
-		}
-		else {
-			map.put("context", String.valueOf(task.getContext()));
-		}
-
-		if (task.getExternalReferenceCode() == null) {
-			map.put("externalReferenceCode", null);
-		}
-		else {
-			map.put(
-				"externalReferenceCode",
-				String.valueOf(task.getExternalReferenceCode()));
-		}
-
-		if (task.getType() == null) {
-			map.put("type", null);
-		}
-		else {
-			map.put("type", String.valueOf(task.getType()));
-		}
-
-		return map;
-	}
-
-	public static class TaskJSONParser extends BaseJSONParser<Task> {
-
-		@Override
-		protected Task createDTO() {
-			return new Task();
-		}
-
-		@Override
-		protected Task[] createDTOArray(int size) {
-			return new Task[size];
-		}
-
-		@Override
-		protected boolean parseMaps(String jsonParserFieldName) {
-			if (Objects.equals(jsonParserFieldName, "context")) {
-				return true;
-			}
-			else if (Objects.equals(
-						jsonParserFieldName, "externalReferenceCode")) {
-
-				return false;
-			}
-			else if (Objects.equals(jsonParserFieldName, "type")) {
-				return false;
-			}
-
-			return false;
-		}
-
-		@Override
-		protected void setField(
-			Task task, String jsonParserFieldName,
-			Object jsonParserFieldValue) {
-
-			if (Objects.equals(jsonParserFieldName, "context")) {
-				if (jsonParserFieldValue != null) {
-					task.setContext((Map<String, ?>)jsonParserFieldValue);
-				}
-			}
-			else if (Objects.equals(
-						jsonParserFieldName, "externalReferenceCode")) {
-
-				if (jsonParserFieldValue != null) {
-					task.setExternalReferenceCode((String)jsonParserFieldValue);
-				}
-			}
-			else if (Objects.equals(jsonParserFieldName, "type")) {
-				if (jsonParserFieldValue != null) {
-					task.setType((String)jsonParserFieldValue);
-				}
-			}
-		}
-
-	}
-
-	private static String _escape(Object object) {
-		String string = String.valueOf(object);
-
-		for (String[] strings : BaseJSONParser.JSON_ESCAPE_STRINGS) {
-			string = string.replace(strings[0], strings[1]);
-		}
-
-		return string;
-	}
-
-	private static String _toJSON(Map<String, ?> map) {
-		StringBuilder sb = new StringBuilder("{");
-
-		@SuppressWarnings("unchecked")
-		Set set = map.entrySet();
-
-		@SuppressWarnings("unchecked")
-		Iterator<Map.Entry<String, ?>> iterator = set.iterator();
-
-		while (iterator.hasNext()) {
-			Map.Entry<String, ?> entry = iterator.next();
-
-			sb.append("\"");
-			sb.append(entry.getKey());
-			sb.append("\": ");
-
-			Object value = entry.getValue();
-
-			sb.append(_toJSON(value));
-
-			if (iterator.hasNext()) {
-				sb.append(", ");
-			}
-		}
-
-		sb.append("}");
-
-		return sb.toString();
-	}
-
-	private static String _toJSON(Object value) {
-		if (value == null) {
-			return "null";
-		}
-
-		if (value instanceof Map) {
-			return _toJSON((Map)value);
-		}
-
-		Class<?> clazz = value.getClass();
-
-		if (clazz.isArray()) {
-			StringBuilder sb = new StringBuilder("[");
-
-			Object[] values = (Object[])value;
-
-			for (int i = 0; i < values.length; i++) {
-				sb.append(_toJSON(values[i]));
-
-				if ((i + 1) < values.length) {
-					sb.append(", ");
-				}
-			}
-
-			sb.append("]");
-
-			return sb.toString();
-		}
-
-		if (value instanceof String) {
-			return "\"" + _escape(value) + "\"";
-		}
-
-		return String.valueOf(value);
-	}
-
-}
